requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4.post0
asyncpraw==7.5.0
discord.py==2.0.1
pyimgur==0.6.0
python-dotenv==0.20.0
SQLAlchemy==1.4.29
